+++ src/main/java/com/senderman/lastkatkabot/repository/AdminUserRepository.java
@@ -3,5 +3,5 @@ package com.senderman.lastkatkabot.repository;
 import com.senderman.lastkatkabot.model.AdminUser;
 import org.springframework.data.repository.CrudRepository;
 
-public interface AdminUserRepository extends CrudRepository<AdminUser, Integer> {
+public interface AdminUserRepository extends CrudRepository<AdminUser, Long> {
 }
