+++ src/main/java/com/senderman/lastkatkabot/repository/BlacklistedUserRepository.java
@@ -3,5 +3,5 @@ package com.senderman.lastkatkabot.repository;
 import com.senderman.lastkatkabot.model.BlacklistedUser;
 import org.springframework.data.repository.CrudRepository;
 
-public interface BlacklistedUserRepository extends CrudRepository<BlacklistedUser, Integer> {
+public interface BlacklistedUserRepository extends CrudRepository<BlacklistedUser, Long> {
 }
