+++ src/main/java/com/senderman/lastkatkabot/service/UserActivityTrackerService.java
@@ -1,5 +1,5 @@
 package com.senderman.lastkatkabot.service;
 
 public interface UserActivityTrackerService {
-    void updateLastMessageDate(long chatId, int userId, int messageLastDate);
+    void updateLastMessageDate(long chatId, long userId, int messageLastDate);
 }
